LayoutTests/platform/wk2/http/tests/loading/basic-auth-load-URL-with-consecutive-slashes-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "frame" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Responding with webkit:rocks
frame "frame" - didCommitLoadForFrame
frame "frame" - didFinishDocumentLoadForFrame
frame "frame" - willPerformClientRedirectToURL: http://127.0.0.1:8000/a//b/non-existent-file.html 
frame "frame" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "frame" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
frame "frame" - didStartProvisionalLoadForFrame
frame "frame" - didCancelClientRedirectForFrame
frame "frame" - didCommitLoadForFrame
frame "frame" - didReceiveTitle: 404 Not Found
frame "frame" - didFinishDocumentLoadForFrame
frame "frame" - didFailLoadWithError
PASS did not cause assertion failure.
